Update the call to get the measures available from a service
The get_measurements call has been updated. Due to a mistake understanding
the inherited code from XiFi, the call to the federation monitoring was done
badly when requirin the data from a VM. This caused a non authorized exception.

Previously, the call to the federation monitoring adapter required the
authorization data as headers, instead of using the data available in the
session. The call to the fereration monitoring adapter has been unified
and the only difference when requiring host or vm data is the url.