Take care on Real Time KPIs when creating agreements
The method that gathers all the KPI info when creating a template
tried to parse the policy without taking into account the possible
value "Real time"

When parsing the value of the policy in the create_agreement function
in the views.py file, we didn't took into account the possible "Real
time" value. Then, we tried to match with a regular expression to get
the hours in the policy. Since it does not match the regular expression
an exception was raised.

Now we check if the value is "Real time" before trying to match the
regular expression.